+++ gui/Settings.h
@@ -20,22 +20,16 @@ enum class RendererEnum {
     /// Surfaces of bodies are meshed using Marching cubes and drawed as triangles.
     MESH,
 
-    /// Raytracer that computes intersections with implicit surface.
-    RAYTRACER,
+    /// Raymarcher that computes intersections with implicit surface.
+    RAYMARCHER,
+
+    /// Volumetric renderer
+    VOLUME,
 
     /// Draws contours (iso-lines) of quantities
     CONTOUR,
 
 };
-static RegisterEnum<RendererEnum> sRenderer({
-    { RendererEnum::NONE, "none", "No particle visualization" },
-    { RendererEnum::PARTICLE, "particle", "Particles are visualized as circles. No shading." },
-    { RendererEnum::MESH,
-        "mesh",
-        "Surfaces of bodies are meshed using Marching cubes and drawed as triangles." },
-    { RendererEnum::RAYTRACER, "raytracer", "Use raytracing to find intersections with implicit surface." },
-    { RendererEnum::CONTOUR, "contour", "Draws contours (iso-lines) of quantities" },
-});
 
 enum class CameraEnum {
     ORTHO,
@@ -43,20 +37,6 @@ enum class CameraEnum {
     FISHEYE,
     SPHERICAL,
 };
-static RegisterEnum<CameraEnum> sCamera({
-    { CameraEnum::ORTHO, "ortho", "Orthographic projection" },
-    { CameraEnum::PERSPECTIVE, "perspective", "Perspective projection" },
-    { CameraEnum::FISHEYE, "fisheye", "Fisheye equidistant projection" },
-    { CameraEnum::SPHERICAL, "spherical", "Spherical 360° projection" },
-});
-
-enum class OrthoEnum { XY, XZ, YZ };
-/// \todo replace with up and dir
-static RegisterEnum<OrthoEnum> sOrtho({
-    { OrthoEnum::XY, "xy", "XY plane" },
-    { OrthoEnum::XZ, "xz", "XZ plane" },
-    { OrthoEnum::YZ, "yz", "YZ plane" },
-});
 
 enum class PlotEnum {
     /// Evolution of the total internal energy in time
@@ -81,42 +61,17 @@ enum class PlotEnum {
     ALL = INTERNAL_ENERGY | KINETIC_ENERGY | TOTAL_ENERGY | TOTAL_MOMENTUM | TOTAL_ANGULAR_MOMENTUM |
           SELECTED_PARTICLE,
 };
-static RegisterEnum<PlotEnum> sPlot({
-    { PlotEnum::INTERNAL_ENERGY, "internal_energy", "Plots the total internal energy." },
-    { PlotEnum::KINETIC_ENERGY, "kinetic_energy", "Plots the total kinetic energy." },
-    { PlotEnum::TOTAL_ENERGY, "total_energy", "Plots the sum of the internal and kinetic energy." },
-    { PlotEnum::TOTAL_MOMENTUM, "total_momentum", "Plots the total momentum." },
-    { PlotEnum::TOTAL_ANGULAR_MOMENTUM, "total_angular_momentum", "Plots the total angular momentum." },
-    /* { PlotEnum::PARTICLE_SFD,
-         "particle_sfd",
-         "Current cumulative size-frequency distribution of bodies in the simulation." },
-     { PlotEnum::PREDICTED_SFD,
-         "predicted_sfd",
-         "Size-frequency distribution that would be realized if we merged all particles that are currently "
-         "gravitationally bounded. It allows to roughly predict the final SFD after reaccumulation. Useful "
-         "for both NBody solvers and SPH solvers." },
-     { PlotEnum::CURRENT_SFD,
-         "current_sfd",
-         "Size-frequency distribution of particle components, i.e. groups of particles in mutual contact. "
-         "Useful for both NBody solvers and SPH solvers. Note that construcing the SFD has non-negligible "
-         "overhead, so it is recommended to specify plot period significantly larger than the time step." },*/
-    { PlotEnum::SELECTED_PARTICLE,
-        "selected_particle",
-        "Plots the current quantity of the selected particle." },
-});
-
-/// \todo generic ortho projection (x,y,z) -> (u,v)
 
 enum class BrdfEnum {
     LAMBERT,
     PHONG,
 };
 
-static RegisterEnum<BrdfEnum> sBrdf({
-    { BrdfEnum::LAMBERT, "lambert", "Lambert shading" },
-    { BrdfEnum::PHONG, "phong", "Phong shading" },
-});
-
+enum class ColorMapEnum {
+    LINEAR,
+    LOGARITHMIC,
+    FILMIC,
+};
 
 enum class GuiSettingsId {
     /// Selected renderer
@@ -162,6 +117,8 @@ enum class GuiSettingsId {
 
     BACKGROUND_COLOR,
 
+    COLORMAP,
+
     SHOW_KEY,
 
     FORCE_GRAYSCALE,
@@ -204,6 +161,8 @@ enum class GuiSettingsId {
     RAYTRACE_BRDF,
 
     RAYTRACE_SPHERES,
+
+    VOLUME_EMISSION,
 
     CONTOUR_SPACING,